[OPIK-375] Introduce `traces` and `LLM calls` tabs in traces table
