+++ apps/opik-frontend/src/components/pages/TracesPage/NoTracesPage.tsx
@@ -6,23 +6,11 @@ import { buildDocsUrl } from "@/lib/utils";
 import { Link } from "@tanstack/react-router";
 import useAppStore from "@/store/AppStore";
 
-type NoTracesPageProps = {
-  name: string;
-};
-
-const NoTracesPage: React.FC<NoTracesPageProps> = ({ name }) => {
+const NoTracesPage = () => {
   const workspaceName = useAppStore((state) => state.activeWorkspaceName);
 
   return (
     <div className="min-w-[340px] py-6">
-      <div className="pb-6">
-        <h1
-          data-testid="traces-page-title"
-          className="comet-title-l truncate break-words"
-        >
-          {name}
-        </h1>
-      </div>
       <div className="flex flex-col items-center rounded-md border bg-white px-6 pb-6 pt-20">
         <h2 className="comet-title-m">Log your first trace</h2>
         <div className="comet-body-s max-w-[570px] px-4 pb-8 pt-4 text-center text-muted-slate">
